requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
contextlib2==21.6.0
Django==5.1
django-cors-headers==4.4.0
django-ninja==1.2.1
django-ninja-extra==0.21.1
django-stubs==5.0.2
django-stubs-ext==5.0.2
dnspython==2.6.1
idna==3.7
injector==0.22.0
mutagen==1.47.0
mysqlclient==2.2.4
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyJWT==2.10.1
PyYAML==6.0.1
requests==2.32.3
sqlparse==0.5.0
types-PyYAML==6.0.12.20240311
typing_extensions==4.12.2
urllib3==2.2.2
websockets==13.0
yt-dlp==2025.7.21
